Parallaxhintergrund hinzufuegen!

+++ Hintergrund.java
@@ -6,14 +6,14 @@ import java.awt.Graphics;
 public class Hintergrund {
     
     private BufferedImage originalBild;
-    private int geschwindigkeit;
+    private double geschwindigkeit;
     
     private int fensterBreite;
     private int fensterHoehe;
     
-    private List<Paar<Integer, BufferedImage>> instanzen;
+    private List<Paar<Double, BufferedImage>> instanzen;
     
-    public Hintergrund(BufferedImage ressource, int geschwindigkeit, int breite, int hoehe) {
+    public Hintergrund(BufferedImage ressource, double geschwindigkeit, int breite, int hoehe) {
         this.originalBild = ressource;
         this.geschwindigkeit = geschwindigkeit;
         this.instanzen = new ArrayList<>();
@@ -27,17 +27,17 @@ public class Hintergrund {
         this.fensterHoehe = hoehe;
         
         if (this.instanzen.size() < 2) {
-            instanzen.add(new Paar<>(0, originalBild));
-            instanzen.add(new Paar<>(this.fensterBreite, originalBild));
+            instanzen.add(new Paar<>(0.0, originalBild));
+            instanzen.add(new Paar<>((double) this.fensterBreite, originalBild));
         }
         
         for (int i = 0; i < instanzen.size(); i ++) {
-            Paar<Integer, BufferedImage> instanz = instanzen.get(i);
+            Paar<Double, BufferedImage> instanz = instanzen.get(i);
             instanz.key += this.geschwindigkeit;
             
             // Unsichtbare Bilder sollen entfernt und mit neuen ersertzt werden!
             if (instanz.key + breite < 0) {
-                instanzen.add(new Paar<>(this.fensterBreite, originalBild));
+                instanzen.add(new Paar<>((double) this.fensterBreite, originalBild));
                 instanzen.remove(i);
             }
         }
@@ -45,8 +45,8 @@ public class Hintergrund {
     
     public void zeichnen(Graphics grafik) {
         for (int i = 0; i < instanzen.size(); i ++) {
-            Paar<Integer, BufferedImage> instanz = instanzen.get(i);
-            grafik.drawImage(instanz.value, instanz.key, 0, fensterBreite, fensterHoehe, null);
+            Paar<Double, BufferedImage> instanz = instanzen.get(i);
+            grafik.drawImage(instanz.value, instanz.key.intValue(), 0, fensterBreite, fensterHoehe, null);
         }
     }
